third_party/WebKit/LayoutTests/dom/node/normalize-with-cdata-expected.txt

Test of normalize on an XML document with CDATA.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Before normalize
PASS serializer.serializeToString(xmlChunk) is "<foo>This is some text before the CDATA<![CDATA[This is some <bold>markup</bold> inside of a CDATA]]>This is some text after the CDATA</foo>"
PASS xmlChunk.documentElement.childNodes.length is 3
After normalize
PASS serializer.serializeToString(xmlChunk) is "<foo>This is some text before the CDATA<![CDATA[This is some <bold>markup</bold> inside of a CDATA]]>This is some text after the CDATA</foo>"
PASS xmlChunk.documentElement.childNodes.length is 3
PASS successfullyParsed is true

TEST COMPLETE
